Refactor skip_if_no_onnxruntime helper to new file
